+++ migrations/20230106190403_add_playlist_id_to_games.sql
@@ -1,0 +1,4 @@
+ALTER TABLE games
+ADD COLUMN playlist_id INTEGER REFERENCES romfiles(id) ON DELETE SET NULL;
+
+CREATE INDEX IF NOT EXISTS games_playlist_id ON games (playlist_id);